git-svn-id: svn://svn.chromium.org/chrome/branches/chrome_official_branch/src@7715 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/history/query_parser.h
@@ -21,7 +21,7 @@ struct QueryWord {
   std::wstring word;
 
   // The starting position of the word in the original text.
-  size_t position;
+  int position;
 };
 
 // QueryNode is used by QueryNodeParser to represent the elements that
